% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district}
\alias{district}
\title{Districts of Tanzania}
\format{
A SpatialPolygonsDataFrame with 195 features:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{Region_Nam} \tab Region name \cr
   \emph{Region_Cod} \tab Region code \cr
   \emph{District_C} \tab District code \cr
   \emph{NewDist20} \tab District name (including new districts) \cr
}
}
\source{
https://www.nbs.go.tz/index.php/en/census-surveys/gis/568-tanzania-districts-shapefiles-2019
}
\usage{
district
}
\description{
Districts of Tanzania
}
\examples{
district

}
\keyword{datasets}
