% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livelihood_zone}
\alias{livelihood_zone}
\title{Livelihoods zones of Tanzania}
\format{
A SpatialPolygonsDataFrame with 80 features:
}
\source{
https://fews.net/east-africa/tanzania/livelihood-zone-map/november-2009
}
\usage{
livelihood_zone
}
\description{
Livelihoods zones of Tanzania
}
\examples{
livelihood_zone

}
\keyword{datasets}
