% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region}
\alias{region}
\title{Regions of Tanzania}
\format{
A SpatialPolygonsDataFrame with 30 features:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{Region_Code} \tab Region code \cr
   \emph{Region_Nam} \tab Region name \cr
}
}
\source{
https://www.nbs.go.tz/index.php/en/census-surveys/gis/384-geographical-information-system-gis
}
\usage{
region
}
\description{
Regions of Tanzania
}
\examples{
region

}
\keyword{datasets}
