% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ruminant_population}
\alias{ruminant_population}
\title{Ruminant population in Tanzania}
\format{
A RasterLayer map
}
\source{
Robinson, Timothy P.; Thornton, Philip; Franceschini, Gianluca;
Kruska, Russ; Chiozza, Federica; Notenbaert, An; Cecchi, Giuliano;
Herrero, Mario; Epprecht, Michael; Fritz, Steffen; You, Liang;
Conchedda, Giulia; See, Linda, 2018, "Global distribution of ruminant
livestock production systems V5 (5 minutes of arc)",
https://doi.org/10.7910/DVN/WPDSZE, Harvard Dataverse, V1
}
\usage{
ruminant_population
}
\description{
Ruminant population in Tanzania
}
\keyword{datasets}
