% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{village}
\alias{village}
\title{Villages of Tanzania}
\format{
A SpatialPolygonsDataFrame with 18421 features:
}
\source{
Tanzania National Bureau of Statistics
}
\usage{
village
}
\description{
Villages of Tanzania
}
\examples{
village

}
\keyword{datasets}
