% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward}
\alias{ward}
\title{Wards of Tanzania}
\format{
A SpatialPolygonsDataFrame with 3644 features:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{Region_Cod} \tab Region code \cr
   \emph{Region_Nam} \tab Region name \cr
   \emph{District_C} \tab District code \cr
   \emph{District_N} \tab District name \cr
   \emph{Ward_Code} \tab Ward code \cr
   \emph{Ward_Name} \tab Ward name \cr
   \emph{Division} \tab Division name \cr
   \emph{SHAPE_Leng} \tab Shape length \cr
   \emph{SHAPE_Area} \tab Shape area \cr
}
}
\source{
https://www.nbs.go.tz/index.php/en/census-surveys/gis/386-2012-phc-shapefiles-level-three
}
\usage{
ward
}
\description{
Wards of Tanzania
}
\examples{
ward

}
\keyword{datasets}
